Add `media_filter` option to `ENMLToHTML` function
Rationale: there are resources that are not images (like pdf, docx, ...) and we don't want to have <IMG src="path/to/document.pdf"/> in the final HTML because it does not have sense for browsers.

This new option is a generic way to tackle this problem. I added the `images_media_filter` function I need as a sample.